requirements.txt
pika==1.1.0
rabbitmq==0.2.0
paho-mqtt==1.5.0
gmqtt==0.6.5